Move code out of URL+Helpers
